重构日志及事件监控，将日志与事件监控相关逻辑分开实现，日志转发及存储目标配置，服务编排系统日志优化

+++ api/src/main/java/cn/jmicro/api/annotation/SMethod.java
@@ -54,7 +54,7 @@ public @interface SMethod {
 	public boolean dumpUpStream() default false;
 	
 	//服务方法级别的日志记录标识，参考monitorEnable说明
-	public int logLevel() default MC.LOG_ERROR;
+	public byte logLevel() default MC.LOG_ERROR;
 	
 	public int retryInterval() default 500;
 	//method must can be retry, or 1
