+++ api/src/main/java/cn/jmicro/api/registry/ServiceMethod.java
@@ -41,7 +41,7 @@ public final class ServiceMethod {
 	
 	//-1 use Service config, 0 disable, 1 enable
 	private int monitorEnable = -1;
-	private int logLevel  = MC.LOG_ERROR;
+	private byte logLevel  = MC.LOG_ERROR;
 	//dump 下行流，用于下行数问题排查
 	private boolean dumpDownStream = false;
 	//dump 上行流，用于上行数问题排查
@@ -296,11 +296,11 @@ public final class ServiceMethod {
 		this.checkInterval = checkInterval;
 	}
 
-	public int getLogLevel() {
+	public byte getLogLevel() {
 		return logLevel;
 	}
 
-	public void setLogLevel(int logLevel) {
+	public void setLogLevel(byte logLevel) {
 		this.logLevel = logLevel;
 	}
